+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/ast/impl/GenericNode.java
@@ -55,11 +55,6 @@ public interface GenericNode<N extends GenericNode<N>> extends Node {
     }
 
     @Override
-    default N getNthParent(int n) {
-        return (N) Node.super.getNthParent(n);
-    }
-
-    @Override
     default NodeStream<N> children() {
         return (NodeStream<N>) Node.super.children();
     }